删除课程、作业时，同时删除文件

+++ api/jeecg-boot-module-system/src/main/java/org/jeecg/modules/teaching/controller/TeachingCourseController.java
@@ -1,41 +1,27 @@
 package org.jeecg.modules.teaching.controller;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-import java.io.IOException;
-import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.jeecg.common.api.vo.Result;
-import org.jeecg.common.aspect.annotation.PermissionData;
-import org.jeecg.common.system.query.QueryGenerator;
-import org.jeecg.common.util.oConvertUtils;
-import org.jeecg.modules.teaching.entity.TeachingCourse;
-import org.jeecg.modules.teaching.service.ITeachingCourseService;
-
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import lombok.extern.slf4j.Slf4j;
-
-import org.jeecgframework.poi.excel.ExcelImportUtil;
-import org.jeecgframework.poi.excel.def.NormalExcelConstants;
-import org.jeecgframework.poi.excel.entity.ExportParams;
-import org.jeecgframework.poi.excel.entity.ImportParams;
-import org.jeecgframework.poi.excel.view.JeecgEntityExcelView;
-import org.jeecg.common.system.base.controller.JeecgController;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.multipart.MultipartHttpServletRequest;
-import org.springframework.web.servlet.ModelAndView;
-import com.alibaba.fastjson.JSON;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import lombok.extern.slf4j.Slf4j;
+import org.jeecg.common.api.vo.Result;
 import org.jeecg.common.aspect.annotation.AutoLog;
+import org.jeecg.common.aspect.annotation.PermissionData;
+import org.jeecg.common.system.base.controller.JeecgController;
+import org.jeecg.common.system.query.QueryGenerator;
+import org.jeecg.modules.system.service.ISysFileService;
+import org.jeecg.modules.teaching.entity.TeachingCourse;
+import org.jeecg.modules.teaching.service.ITeachingCourseService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.ModelAndView;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Arrays;
+import java.util.List;
 
  /**
  * @Description: 课程
@@ -50,6 +36,8 @@ import org.jeecg.common.aspect.annotation.AutoLog;
 public class TeachingCourseController extends JeecgController<TeachingCourse, ITeachingCourseService> {
 	@Autowired
 	private ITeachingCourseService teachingCourseService;
+	@Autowired
+	private ISysFileService sysFileService;
 
 	 @GetMapping("/mineCourse")
 	 public Result<List<TeachingCourse>> mineCourse(){
@@ -118,7 +106,13 @@ public class TeachingCourseController extends JeecgController<TeachingCourse, IT
 	@ApiOperation(value="课程-通过id删除", notes="课程-通过id删除")
 	@DeleteMapping(value = "/delete")
 	public Result<?> delete(@RequestParam(name="id",required=true) String id) {
-		teachingCourseService.removeById(id);
+		TeachingCourse course = this.teachingCourseService.getById(id);
+		if (course != null){
+			sysFileService.deleteWithFile(course.getCourseIcon());
+			sysFileService.deleteWithFile(course.getCourseCover());
+			sysFileService.deleteWithFile(course.getCourseMap());
+			teachingCourseService.removeById(id);
+		}
 		return Result.ok("删除成功!");
 	}
 	
@@ -132,7 +126,14 @@ public class TeachingCourseController extends JeecgController<TeachingCourse, IT
 	@ApiOperation(value="课程-批量删除", notes="课程-批量删除")
 	@DeleteMapping(value = "/deleteBatch")
 	public Result<?> deleteBatch(@RequestParam(name="ids",required=true) String ids) {
-		this.teachingCourseService.removeByIds(Arrays.asList(ids.split(",")));
+		List<String> idList = Arrays.asList(ids.split(","));
+		List<TeachingCourse> courseList = this.teachingCourseService.list(new QueryWrapper<TeachingCourse>().in("id", idList));
+		for (TeachingCourse course: courseList){
+			sysFileService.deleteWithFile(course.getCourseIcon());
+			sysFileService.deleteWithFile(course.getCourseCover());
+			sysFileService.deleteWithFile(course.getCourseMap());
+		}
+		this.teachingCourseService.removeByIds(idList);
 		return Result.ok("批量删除成功!");
 	}
